reorganize profile page folder
