+++ src/typescript/ui/elements/main-content-pages/profile-page/profile-nft-card-element/toniq-entrepot-profile-nft-list-item.element.ts
@@ -1,11 +1,11 @@
 import {classMap} from 'lit/directives/class-map.js';
 import {assign, css, html, renderIf} from 'element-vir';
 import {defineToniqElement, toniqFontStyles, Icp16Icon, ToniqIcon} from '@toniq-labs/design-system';
-import {toIcp} from '../../../../data/icp';
-import {FullProfileNft} from './profile-nfts/full-profile-nft';
+import {toIcp} from '../../../../../data/icp';
+import {FullProfileNft} from '../profile-page-nfts/full-profile-nft';
 import {isRuntimeTypeOf} from '@augment-vir/common';
-import {BaseNft} from '../../../../data/nft/base-nft';
-import {EntrepotNftListItemElement} from '../../common/toniq-entrepot-nft-list-item.element';
+import {BaseNft} from '../../../../../data/nft/base-nft';
+import {EntrepotNftListItemElement} from '../../../common/toniq-entrepot-nft-list-item.element';
 import {EntrepotProfileNftListItemTextItemsElement} from './toniq-entrepot-profile-nft-list-item-text-items.element';
 
 export const EntrepotProfileNftListItemElement = defineToniqElement<{
