+++ src/typescript/ui/elements/main-content-pages/profile-page/profile-page-nfts/create-right-column-template.ts
