+++ src/typescript/ui/elements/main-content-pages/profile-page/profile-page-state/profile-page-nft-parts/nft-profile-offer-status.ts
@@ -1,4 +1,4 @@
-import {BaseNft} from '../../../../../data/nft/base-nft';
+import {BaseNft} from '../../../../../../data/nft/base-nft';
 
 export type ProfileNftOfferStatus = 'none' | 'received' | 'offered' | 'offered received';
