+++ src/typescript/ui/elements/main-content-pages/profile-page/profile-page-state/profile-tabs.ts
@@ -1,5 +1,5 @@
-import {ensureType, mapObjectValues} from '@augment-vir/common';
-import {TopTab} from '../../common/toniq-entrepot-top-tabs.element';
+import {mapObjectValues} from '@augment-vir/common';
+import {TopTab} from '../../../common/toniq-entrepot-top-tabs.element';
 
 const profileTabValueToLabelMap = {
     'my-nfts': 'My NFTs',
